+++ geowatch/cli/smartflow/run_iarpa_metrics.py
@@ -112,6 +112,9 @@ class RunIARPAMetricsCLI(scfg.DataConfig):
         print("* Running baseline framework kwcoco ingress *")
         ingress_dir = ub.Path('/tmp/ingress')
 
+        eval_dpath = (ingress_dir / 'metrics_output').ensuredir()
+        viz_output_dpath = (eval_dpath / 'region_viz_overall').ensuredir()
+
         USE_NON_STAC_PATHS = True
         if USE_NON_STAC_PATHS:
             # FIXME: would be better if we conformed to the same STAC-in
@@ -127,19 +130,20 @@ class RunIARPAMetricsCLI(scfg.DataConfig):
             remote_true_site_fpaths = [p for p in remote_true_site_dpath.ls() if p.name.startswith(config.region_id)]
             remote_true_region_fpaths = [p for p in remote_true_region_dpath.ls() if p.name.startswith(config.region_id)]
 
-            true_annot_dpath = (ingress_dir / 'truth/region_models').ensuredir()
+            true_annot_dpath = (ingress_dir / 'truth').ensuredir()
             true_region_dpath = (true_annot_dpath / 'region_models').ensuredir()
             true_site_dpath = (true_annot_dpath / 'site_models').ensuredir()
 
             # Copy predictions to local node
-            pred_site_dpath = FSPath.coerce(ingress_dir / 'site_models')
+            pred_annot_dpath = (ingress_dir / 'pred').ensuredir()
+            pred_site_dpath = FSPath.coerce(pred_annot_dpath / 'site_models')
             remote_pred_site_dpath.copy(pred_site_dpath, verbose=3)
 
-            if 0:
-                # HACK
-                remote_pred_region_dpath = remote_pred_site_dpath.parent / 'region_models'
-                pred_region_dpath = FSPath.coerce(ingress_dir / 'region_models')
-                remote_pred_region_dpath.copy(pred_region_dpath)
+            # We dont need the region model for scoring, but it helps for
+            # visualization and ensuring our approch is consistent
+            remote_pred_region_dpath = remote_pred_site_dpath.parent / 'region_models'
+            pred_region_dpath = FSPath.coerce(pred_annot_dpath / 'region_models')
+            remote_pred_region_dpath.copy(pred_region_dpath)
 
             # Copy truth to local node
             for fpath in ub.ProgIter(remote_true_site_fpaths, desc='pull site truth', verbose=3):
@@ -147,7 +151,7 @@ class RunIARPAMetricsCLI(scfg.DataConfig):
             for fpath in ub.ProgIter(remote_true_region_fpaths, desc='pull region truth', verbose=3):
                 fpath.copy(true_region_dpath / fpath.name)
 
-        # # 2. Download and prune region file
+        # 2. Download and prune region file
         print("* Downloading and pruning region file *")
         local_region_path = '/tmp/region.json'
         download_region(
@@ -159,9 +163,43 @@ class RunIARPAMetricsCLI(scfg.DataConfig):
 
         node_state.print_current_state(ingress_dir)
 
+        DRAW_SANITY_CHECK = True
+        if DRAW_SANITY_CHECK:
+            print('Draw predictions and truth before we do eval as a sanity check')
+            command = ub.paragraph(
+                f'''
+                geowatch draw_region {true_site_dpath}
+                    --extra_header "True Sites"
+                    --fpath "{viz_output_dpath}"/true_site_viz.png
+                ''')
+            ub.cmd(command, shell=True, verbose=3)
+
+            command = ub.paragraph(
+                f'''
+                geowatch draw_region {true_region_dpath}
+                    --extra_header "True Region"
+                    --fpath "{viz_output_dpath}"/true_region_viz.png
+                ''')
+            ub.cmd(command, shell=True, verbose=3)
+
+            command = ub.paragraph(
+                f'''
+                geowatch draw_region {pred_site_dpath}
+                    --extra_header "Pred Sites"
+                    --fpath "{viz_output_dpath}"/pred_site_viz.png
+                ''')
+            ub.cmd(command, shell=True, verbose=3)
+
+            command = ub.paragraph(
+                f'''
+                geowatch draw_region {pred_region_dpath}
+                    --extra_header "Pred Regions"
+                    --fpath "{viz_output_dpath}"/pred_region_viz.png
+                ''')
+            ub.cmd(command, shell=True, verbose=3)
+
         smart_pipeline.PolygonEvaluation.name = 'poly_eval'
         eval_node = smart_pipeline.PolygonEvaluation()
-        eval_dpath = (ingress_dir / 'metrics_output').ensuredir()
         eval_fpath = eval_dpath / 'poly_eval.json'
 
         eval_node.configure({
@@ -179,6 +217,46 @@ class RunIARPAMetricsCLI(scfg.DataConfig):
         node_state.print_current_state(ingress_dir)
 
         node_state.print_directory_contents(eval_dpath)
+
+        # Change viz symlinks to real files
+        if True:
+            for path in viz_output_dpath.glob('*'):
+                if path.is_symlink():
+                    real_path = path.resolve()
+                    assert not real_path.is_symlink()
+                    path.unlink()
+                    real_path.copy(path)
+
+        if True:
+            confusion_out_dpath = eval_dpath / 'confusion_analysis'
+            region_id = config.region_id
+            bas_metric_dpath = (eval_dpath / region_id / 'overall/bas')
+            src_kwcoco_fpath = None  # TODO: get imagery
+            command = ub.codeblock(
+                fr'''
+                python -m geowatch.mlops.confusor_analysis \
+                    --src_kwcoco="{src_kwcoco_fpath}" \
+                    --bas_metric_dpath="{bas_metric_dpath}" \
+                    --pred_sites="{pred_site_dpath}" \
+                    --out_dpath={confusion_out_dpath} \
+                    --true_region_dpath="{true_region_dpath}" \
+                    --true_site_dpath="{true_site_dpath}" \
+                    --region_id="{region_id}" \
+                    --viz_sites=0 \
+                    --reload=0
+                ''')
+            ub.cmd(command, verbose=3, shell=1)
+
+            confusion_group_dpath = confusion_out_dpath / 'confusion_groups'
+            confusion_region_fpaths = list(confusion_group_dpath.glob('*.geojson'))
+            for fpath in confusion_region_fpaths:
+                command = ub.paragraph(
+                    f'''
+                    geowatch draw_region {fpath}
+                        --extra_header "{fpath.name}"
+                        --fpath "{fpath}.png"
+                    ''')
+                ub.cmd(command, shell=True, verbose=3)
 
         assets_to_egress = {
             'eval_dpath': eval_dpath,