+++ geowatch/tasks/fusion/datamodules/kwcoco_datamodule.py
@@ -105,7 +105,7 @@ class KWCocoVideoDataModuleConfig(KWCocoVideoDatasetConfig):
     sqlview = scfg.Value(False, help=ub.paragraph(
             '''
             If False, reads the COCO dataset as a json file. Otherwise
-            it can be sqlite or postgresql to cache json file in an SQL
+            it can be "sqlite" or "postgresql" to cache json file in an SQL
             database for faster responce times and lower memory
             footprint.
             '''))