+++ scripts/prepare_drop8_ARA.sh
@@ -43,6 +43,9 @@ SITE_GLOBSTR="$DVC_DATA_DPATH/annotations/drop8-v1/site_models/*_*0*_*.geojson"
 #REGION_GLOBSTR="$DVC_DATA_DPATH/annotations/drop8-v1/region_models/KW_C001.geojson"
 #SITE_GLOBSTR="$DVC_DATA_DPATH/annotations/drop8-v1/site_models/KW_C001_*.geojson"
 
+REGION_GLOBSTR="$DVC_DATA_DPATH/annotations/drop8-v1/region_models/KR_R002.geojson"
+SITE_GLOBSTR="$DVC_DATA_DPATH/annotations/drop8-v1/site_models/KR_R002_*.geojson"
+
 
 # iMerit Regions Only
 #REGION_GLOBSTR="$DVC_DATA_DPATH/annotations/drop8-v1/region_models/*_C*.geojson"
@@ -93,7 +96,7 @@ python -m geowatch.cli.queue_cli.prepare_ta2_dataset \
     --cache=$CACHE_STEPS \
     --verbose=100 \
     --skip_existing=0 \
-    --force_min_gsd=2.0 \
+    --force_min_gsd=0.1 \
     --force_nodata=-9999 \
     --align_tries=1 \
     --asset_timeout="10 minutes" \